migrate reused localize constants
