+++ src/commands/createFunction/durableSteps/netherite/NetheriteEventHubNameStep.ts
@@ -7,11 +7,11 @@ import { Eventhub, EventHubManagementClient } from '@azure/arm-eventhub';
 import { parseAzureResourceId, uiUtils } from '@microsoft/vscode-azext-azureutils';
 import { AzureWizardPromptStep, ISubscriptionContext, nonNullValue } from '@microsoft/vscode-azext-utils';
 import { ConnectionType } from '../../../../constants';
-import { invalidLength, invalidLowerCaseAlphanumericWithHyphens, localize } from '../../../../localize';
+import { invalidLength, invalidLowerCaseAlphanumericWithHyphens } from '../../../../constants-nls';
+import { localize } from '../../../../localize';
 import { createEventHubClient } from '../../../../utils/azureClients';
 import { validateUtils } from '../../../../utils/validateUtils';
 import { IEventHubsConnectionWizardContext } from '../../../appSettings/IEventHubsConnectionWizardContext';
-
 
 export class NetheriteEventHubNameStep<T extends IEventHubsConnectionWizardContext> extends AzureWizardPromptStep<T> {
     private _eventHubs: Eventhub[] = [];
