+++ src/commands/createNewProject/NewProjectLanguageStep.ts
@@ -6,7 +6,8 @@
 import { AzureWizardExecuteStep, AzureWizardPromptStep, IAzureQuickPickItem, IWizardOptions, UserCancelledError } from '@microsoft/vscode-azext-utils';
 import { QuickPickOptions } from 'vscode';
 import { previewPythonModel, ProjectLanguage, pysteinModelSetting } from '../../constants';
-import { localize, pythonNewModelPreview } from '../../localize';
+import { pythonNewModelPreview } from '../../constants-nls';
+import { localize } from '../../localize';
 import { nonNullProp } from '../../utils/nonNull';
 import { openUrl } from '../../utils/openUrl';
 import { pythonUtils } from '../../utils/pythonUtils';
