+++ src/utils/venvUtils.ts
@@ -7,8 +7,9 @@ import { AzExtFsExtra } from '@microsoft/vscode-azext-utils';
 import * as path from 'path';
 import * as vscode from 'vscode';
 import { pythonVenvSetting, requirementsFileName } from "../constants";
+import { emptyWorkspace } from '../constants-nls';
 import { ext } from '../extensionVariables';
-import { emptyWorkspace, localize } from '../localize';
+import { localize } from '../localize';
 import { getWorkspaceSetting } from '../vsCodeConfig/settings';
 import { cpUtils } from './cpUtils';
 import { getWorkspaceRootPath } from './workspace';
